refactor: error 디렉토리 및 파일 구성방식 수정
