+++ apps/frontend/src/_libs/error/errors/APIError.ts
@@ -1,0 +1,20 @@
+import AppError, { AppErrorParams } from './index';
+
+type APIErrorParams<TData = unknown> = AppErrorParams<TData> & { status: number };
+
+export class APIError<TData = unknown> extends AppError<TData> {
+  public status: number;
+
+  constructor({ message, status, code, data, reset }: APIErrorParams<TData>) {
+    super({
+      message,
+      code,
+      data,
+      reset,
+    });
+
+    this.name = 'APIError';
+
+    this.status = status;
+  }
+}
